+++ src/assets/package.lisp
@@ -6,5 +6,6 @@
    #:add-asset-root #:remove-asset-root #:resolve-asset
    #:load-text-file #:load-binary-file
    #:register-asset-loader #:unregister-asset-loader #:asset-loader
-   #:load-asset #:invalidate-asset #:clear-asset-cache #:reload-changed-assets
-   #:cached-asset-p #:asset-cache-size))
+   #:load-asset #:preload-assets #:invalidate-asset #:clear-asset-cache #:reload-changed-assets
+   #:cached-asset-p #:asset-cache-size #:cached-assets
+   #:add-asset-reload-listener #:remove-asset-reload-listener))